Rename dweb components from .js to .jsx also means changing all imports to explicitly specify .jsx as webpack not picking up the automatic extension
