+++ packages/ia-components/sandbox/details/DetailsFlags.jsx
