+++ packages/ia-components/sandbox/details/ModalGo.jsx
