+++ packages/ia-components/sandbox/tiles/TileImage.jsx
